+++ src/array.mjs
@@ -63,3 +63,4 @@ export { default as sort } from './array/sort.mjs';
 export { default as array_multisort } from './array/array_multisort.mjs';
 export { default as rsort } from './array/rsort.mjs';
 export { default as usort } from './array/usort.mjs';
+export { default as uksort } from './array/uksort.mjs';