+++ tests/array.global.test.mjs
@@ -67,4 +67,5 @@ test('array.global', () => {
     expect(array_multisort).not.toBeUndefined();
     expect(rsort).not.toBeUndefined();
     expect(usort).not.toBeUndefined();
+    expect(uksort).not.toBeUndefined();
 });